requirements.txt
autopep8==2.0.1
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
idna==3.4
mypy-extensions==1.0.0
mysql-connector-python==8.0.32
ory-client==1.1.25
packaging==23.0
pathspec==0.11.0
pip-autoremove==0.10.0
platformdirs==3.0.0
protobuf==3.20.3
psycopg2-binary==2.9.5
pycodestyle==2.10.0
python-dateutil==2.8.2
python-dotenv==0.21.1
requests==2.29.0
six==1.16.0
tomli==2.0.1
typing-extensions==4.5.0
urllib3==1.26.15
